+++ people-mgmt-service/src/main/java/com/jmk/people/entity/Person.java
@@ -88,7 +88,7 @@ public class Person {
 	private Integer groupId = null;
 
 	@Column(name = "created_by")
-	private Integer createdBy = null;
+	private Long createdBy = null;
 
 	@Column(name = "created_on")
 	private Timestamp createdOn = null;
@@ -326,14 +326,14 @@ public class Person {
 	/**
 	 * @return the createdBy
 	 */
-	public Integer getCreatedBy() {
+	public Long getCreatedBy() {
 		return createdBy;
 	}
 
 	/**
 	 * @param createdBy the createdBy to set
 	 */
-	public void setCreatedBy(Integer createdBy) {
+	public void setCreatedBy(Long createdBy) {
 		this.createdBy = createdBy;
 	}
